Update 1.copy_ids.sh
This is the first step in the analysis. All genotype file (in .zip) format are saved in /src folder. We should have a file called "lambids" for which we want to extract genotype data for GWAS analysis. This piece of code will extract genotype files for all the lamb ids to be analysed and save it to destination specified.